+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.4.0] - 2018-10-03
+
+### Added
+- use [mstime](https://github.com/ngduc/mstime) to measure API run time.
+### Changed
+- BREAKING: refactor apiJson's "listModel" to "model"
+
 ## [0.3.0] - 2018-10-02
+
 ### Changed
 - BREAKING: refactor code to use this syntax: import { User } from 'api/models';